Droplets

Superdroplet Method Logic for Cloud Microphysics

Functions

Droplets.FK — Method

FK(T)
FD(T)
Functions in the Denominator of the Köhler Equation

Uses the constants structs defined in Droplets

Arguments

T: Temperature in K

Droplets.adaptive_pαdt! — Method

adaptive_pαdt!(L::Vector{Tuple{Int,Int}}, droplets::droplet_attributes, coag_data::coagulation_run, t_left::Ref{FT}, kernel::Function, settings::coag_settings{FT}) where FT<:AbstractFloat

Map the probability function over the list of droplet pairs, L, using adaptive timestepping logic and update the coagulation data in place.

Arguments

L::Vector{Tuple{Int,Int}}: List of droplet indices.

droplets::droplet_attributes: Droplet attributes.

coag_data::coagulation_run: Coagulation data.

t_left::Ref{FT}: Timestep after adaptive substep.

kernel::Function: Coalescence kernel function.

settings::coag_settings{FT}: Coagulation settings.

Droplets.calc_ρw — Method

calc_ρw(ξ, X, ΔV)

Calculate the density of liquid water (ρw) based on the given droplet multiplicity(ξ) and volume(X), and domain volume.

Returns

ρw: density of water

Droplets.calc_ρw — Method

calc_ρw(droplets, ΔV)

Calculate the density of water (ρw) based on the given droplet attributes and domain volume.

Arguments

droplets: an array of droplets with attributes ξ (multiplicity) and X (volume).

ΔV: grid cell volume.

Returns

ρw: density of water

Droplets.coalescence_timestep! — Method

coalescence_timestep!(run::backend, scheme::scheme_type, droplets::droplet_attributes)

Perform a coalescence timestep for the given droplets using the Superdroplet Method (SDM) Shima et al. (2009) when the lowest multiplicity of superdroplets is less than 1, the largest superdroplet is split into two equal parts, as proposed by Dziekan and Pawlowska (ACP, 2017) https://doi.org/10.5194/acp-17-13509-2017

Arguments

run::backend: Threading over Linear Sampling option

scheme::schemetype: adaptive or none

droplets::droplet_attributes: The superdroplets.

Droplets.compute_pαdt! — Method

compute_pαdt!(L::Vector{Tuple{Int,Int}}, droplets::droplet_attributes, coag_data::coagulation_run, kernel::Function, scale::FT,scalecoagsettings::coag_settings{FT}) where FT<:AbstractFloat

Map the probability function over the list of droplet pairs, L, and update the coagulation data in place.

Arguments

L::Vector{Tuple{Int,Int}}: List of droplet indices to be considered for coalescence.

droplets::droplet_attributes: Droplet attributes.

coag_data::coagulation_run: Coagulation data.

kernel::Function: Coalescence kernel function.

scale::FT: Scaling factor for the linear coalescence probability.

coagsettings::coag_settings{FT}: Coagulation settings.

Droplets.dXkohler_function_of_radius — Method

dXkohler_function_of_radius(R, M, m, T, Senv, timestep)

Calculate the change in droplet volume due to condensation using the Kohler equation.

Arguments

R: Droplet radius

M: Molecular weight of the droplet substance

m: Molecular weight of the dry air

T: Temperature

Senv: Saturation of the environment

timestep: Time step

Returns

dX: Change in droplet mass

Droplets.dXkohler_function_of_radius — Method

dXkohler_function_of_radius(R, M, m, T, qv, P, timestep)

Calculate the change in droplet volume due to condensation using the Kohler equation.

Arguments

R: Droplet radius

M: Molecular weight of the droplet substance

m: Molecular weight of the dry air

T: Temperature

qv: Water vapor mixing ratio

P: Pressure

timestep: Time step

Returns

dX: Change in droplet volume

Droplets.dXkohler_function_of_radius_activated — Method

dXkohler_function_of_radius_activated(R, T, Senv, timestep)

Calculate the change in droplet volume due to condensation using the Kohler equation for activated droplets, neglecting solute.

Arguments

R: Droplet radius

T: Temperature

Senv: Saturation of the environment

timestep: Time step

Returns

dX: Change in droplet mass

Droplets.dq_liq_cond — Method

dq_liq_cond(R, M, m, T, Senv, timestep, ρ_air)

Calculate the change in q, liquid mixing ratio due to condensation of droplets, using droplet solute information.

Arguments

R: Droplet radius, meters

M: Molecular weight of the droplet

m: Molecular weight of the solute

T: Temperature

Senv: Environmental saturation

timestep: Time step

ρ_air: Density of air

Returns

dql: Change in liquid water mass

Droplets.dq_liq_cond_activated — Method

dq_liq_cond_activated(R, T, Senv, timestep, ρ_air)
dq_liq_cond_activated(R, M, m, T, Senv, timestep, ρ_air)

Calculate the change in liquid water mass due to condensation of activated droplets.

Arguments

R: Droplet radius, meters

T: Temperature

Senv: Environmental saturation

timestep: Time step

ρ_air: Density of air

Returns

dql: Change in liquid water mass

Droplets.drkohler — Method

Methods: -drkohler(R, M, m, T, qv, P, timestep) -drkohler(R, M, m, T, Senv, timestep)

RHS of the Köhler equation.

Arguments

R: Droplet radius (m)

M: Molecular weight of the solute (kg/mol)

m: Molar mass of the solute (g/mol)

T: Temperature (K)

qv: Water vapor mixing ratio (kg/kg)

P: Atmospheric pressure (Pa)

Senv: Environmental saturation (dimensionless)

timestep: Time step (s)

Returns

dr: Change in droplet radius over the timestep (m)

Droplets.drkohler_activated — Method

drkohler_activated(R, T, Senv, timestep)

Compute the rate of change of droplet radius for activated droplets, neglecting differences in solutes.

Arguments

R: Radius of the droplets.

T: Temperature.

Senv: Environmental supersaturation.

timestep: simulation Time step.

Returns

dr: rate of change of droplet radius.

Droplets.golovin — Method

golovin(droplets, (j, k), settings)

Golovin (or additive) kernel function for droplet coalescence, used to calculate the probabilities of two droplets colliding. K(x,x') = b(x+x'),where b is the Golovin kernel coefficient, and x is the volume of the droplet. Taken from Golovin (1963), this kernel has an analytic solution for the Smulochowski Coagulation Equation.

Arguments

droplets: droplet attributes

(j, k): indices of the droplets

settings: coagulation settings

Droplets.hydrodynamic — Method

hydrodynamic(droplets::droplet_attributes, (j,k)::Tuple{Int,Int}, settings::coag_settings{FT})::FT where FT<:AbstractFloat

Hydronynamic kernel function for droplet coalescence, used to calculate the probabilities of two droplets colliding. K(r,r') = π * (r+r')^2 * |v(r)-v(r')| * E(r,r') where E(r,r') is the collision efficiency function, (currently not implemented), and v is the terminal velocity of a droplet of radius size r.

Arguments

droplets::droplet_attributes: The attributes of the droplets.

(j,k)::Tuple{Int,Int}: The indices of the droplets.

settings::coag_settings{FT}: The coagulation settings.

Droplets.init_ξ_const — Method

init_ξ_const(settings::coag_settings{FT}) where FT<:AbstractFloat

initξconst initializes droplets based on the constant-multiplicity initialization method, using an exponential distribution around the initial volume of the droplets. The method is as described in Shima et al. (2009) https://doi.org/10.5194/acp-9-4491-2009

Arguments

settings: Coagulation settings, type ::coag_settings.

Droplets.pair_Ps_adaptive! — Method

pair_Ps_adaptive!(α::Int, pair::Tuple{Int,Int}, droplets::droplet_attributes, coag_data::coagulation_run, t_max::Vector{FT})

This function calculates the coalescence probability for a given pair of droplets, using adaptive timestepping logic (Bartman et al. 2021). It takes the following arguments:

α::Int: The index of the coalescence model.

pair::Tuple{Int,Int}: The indices of the droplets in the pair.

droplets::droplet_attributes: The attributes of the droplets.

coag_data::coagulation_run: The coagulation run data.

t_max::Vector{FT}: The maximum timestep allowed given multiple sampling.

Droplets.radius_to_volume — Method

radius_to_volume(R::AbstractFloat)

Volume of sphere with radius R

Arguments

R::AbstractFloat: Radius in meters.

Returns

The volume of the droplet in meters^3.

Droplets.terminal_v — Method

terminal_v(r::FT) where FT<:AbstractFloat

Compute the terminal velocity of a droplet of radius r, using tables from Gunn and Kinzer, (1949), https://doi.org/10.1175/1520-0469(1949)006<0243:TTVOFF>2.0.CO;2

Arguments

r::FT: The radius of the droplet in meters.

Returns

The terminal velocity of the droplet in meters/second.

Droplets.volume_to_radius — Method

volume_to_radius(V::AbstractFloat)

radius based on spherical volume

Arguments

V::AbstractFloat: Spherical volume in meters^3.

Returns

The radius of the droplet in meters.

Types

Droplets.Constants — Type

Constants{FT<:AbstractFloat}

A struct representing constants used in the Droplets package.

Fields

FT: The type of the constants, which must be a subtype of AbstractFloat.

k: thermal conductivity (default: 0.024).

μ: dynamic viscosity (default: 1.6e-5).

Dv: diffusion coefficient of water vapor in air (default: 2.26e-5).

ρl: density of liquid water (default: 1000.0).

Rd: gas constant for dry air (default: 287.0).

Rv: gas constant for water vapor (default: 461.0).

gconst: gravitational constant (default: 9.8).

L: latent heat of vaporization (default: 22.6e5).

Cp: specific heat of dry air at constant pressure (default: 4181).

Droplets.coag_settings — Type

coag_settings{FT<:AbstractFloat}

A struct representing the settings for coalescence.

Fields

FT: AbstractFloat.

Δt: The time step for the coalescence simulation.

ΔV: The volume of the coalescence domain.

Ns: The number of superdroplets.

scale: The scaling factor that allows for linear sampling.

R_min: The minimum radius of sampled droplets, used in some initializations. Radius in meters

R_max: The maximum radius of sampled droplets, used in some initializations. Radius in meters.

golovin_kernel_coeff: The Golovin kernel coefficient, 1/seconds.

hydrodynamic_collision_eff_func: A boolean indicating whether to use the hydrodynamic collision efficiency function. Currently not implemented.

kernel: The kernel function used for coalescence. Implemented options are golovin or hydrodynamic, can take any user implemented function with input type (::dropletattributes, pairindex::Tuple{Int,Int}, ::coagsettings{FT}).

n0: The initial real world droplet concentration.

R0: The initial seed radius of the droplets.

Droplets.coagulation_run — Type

struct coagulation_run{FT<:AbstractFloat}

Struct initialising temp memory used for coagulation.

Input:

Ns::Int : Number of superdroplets.

Fields:

I::Vector{Int} : Vector of indexes to be shuffled for permutations.

pαdt::Vector{FT} : Vector of coalescence probabilities for each pair.

ϕ::Vector{FT} : Random numbers to be used in Monte Carlo coalescence.

lowest_zero::Ref{Bool} : Reference to a boolean indicating if the lowest multiplicity is zero.

deficit::Ref{FT} : Reference to a float representing the deficit in mass or volume.

Droplets.droplet_attributes — Type

struct droplet_attributes{FT<:AbstractFloat}

A struct representing the attributes of a droplet.

Fields:

FT: AbstractFloat.

ξ: Vector of integers representing the multiplicity of each droplet.

R: Vector of floats representing the radius of each droplet.

X: Vector of floats representing the volume of each droplet.

Droplets.run_settings — Type

struct run_settings{FT<:AbstractFloat}

A struct representing the run settings for the binning process.

Fields:

FT: The type of floating-point numbers used in the run settings.

num_bins: The number of bins to use for binning output(default: 128).

radius_bins_edges: The edges of the bins for the radius (default: logarithmic scale from 10e-6 to 5e3e-6).

smooth: A boolean indicating whether to apply smoothing to the bins (default: true).

smooth_scope: The scope of smoothing (default: 2).

init_random_seed: The random seed for initialization (default: 30).

coag_threading: The threading method for coalescence (default: Serial).

Serial() or Parallel()

scheme: The scheme to use for coalescence (default: none). -none(), Adaptive()

output_steps: The time steps at which to output the results (default: [0, 1200, 2400, 3600]).

init_method: The method to use for initialization (default: init_logarithmic).

initlogarithmic,initξconst, inituniform_sd

binning_method: The method to use for binning (default: massdensitylnr). (a function that takes droplets and coagsettings as arguments)

normalize_bins_dlnr: A boolean indicating whether to normalize the bins by the logarithm of the radius (default: true).