Constants

Droplets.Constants — Type

Constants{FT<:AbstractFloat}

A struct representing constants used in the Droplets package.

Fields

FT: The type of the constants, which must be a subtype of AbstractFloat.

k: thermal conductivity (default: 0.024).

μ: dynamic viscosity (default: 1.6e-5).

Dv: diffusion coefficient of water vapor in air (default: 2.26e-5).

ρl: density of liquid water (default: 1000.0).

Rd: gas constant for dry air (default: 287.0).

Rv: gas constant for water vapor (default: 461.0).

gconst: gravitational constant (default: 9.8).

L: latent heat of vaporization (default: 22.6e5).

Cp: specific heat of dry air at constant pressure (default: 4181).

σSB: Stefan-Boltzmann constant (default: 5.670374419e-8).

T0: reference temperature (default: 273.15).

P0: reference pressure (default: 100000.0).

P_SLP: standard sea level pressure (default: 101325.0).

ϵ: molar mass ratio of water vapor to dry air (default: 0.622).

Ω: Earth's rotation rate in radians per second (default: 7.2921150e-5).

dry_theta_convention: boolean indicating whether to use dry partial pressure convention for potential temperature (default: false), build for prescribed KiD thermo.

Droplets.radius_to_volume — Method

radius_to_volume(R::AbstractFloat)

Volume of sphere with radius R

Arguments

R::AbstractFloat: Radius in meters.

Returns

The volume of the droplet in meters^3.

Droplets.volume_to_radius — Method

volume_to_radius(V::AbstractFloat)

radius based on spherical volume

Arguments

V::AbstractFloat: Spherical volume in meters^3.

Returns

The radius of the droplet in meters.