#Collision Kernels

Droplets._tv_d_table — Constant

terminal_v(r::FT) where FT<:AbstractFloat

Compute the terminal velocity of a droplet of radius r, using tables from Gunn and Kinzer, (1949), https://doi.org/10.1175/1520-0469(1949)006<0243:TTVOFF>2.0.CO;2

Arguments

r::FT: The radius of the droplet in meters.

Returns

The terminal velocity of the droplet in meters/second.

Droplets.collision_efficiency — Method

collision_efficiency(R1::FT, R2::FT) where FT<:AbstractFloat

Compute the collision efficiency between two droplets of radius R1 and R2 using the Hall-Davis parameterization.

Droplets.golovin — Method

golovin(droplets, (j, k), settings)

Golovin (or additive) kernel function for droplet coalescence, used to calculate the probabilities of two droplets colliding. K(x,x') = b(x+x'),where b is the Golovin kernel coefficient, and x is the volume of the droplet. Taken from Golovin (1963), this kernel has an analytic solution for the Smulochowski Coagulation Equation.

Arguments

droplets: droplet attributes

(j, k): indices of the droplets

settings: coagulation settings

Droplets.hall_davis_idx — Method

Compute the index into the Hall-Davis collision efficiency matrix for a droplet of radius R.

Droplets.hydrodynamic — Method

hydrodynamic(droplets::droplet_attributes, (j,k)::Tuple{Int,Int}, settings::coag_settings{FT})::FT where FT<:AbstractFloat

Hydronynamic kernel function for droplet coalescence, used to calculate the probabilities of two droplets colliding. K(r,r') = π * (r+r')^2 * |v(r)-v(r')| * E(r,r') where E(r,r') is the collision efficiency function, (currently not implemented), and v is the terminal velocity of a droplet of radius size r.

Arguments

droplets::droplet_attributes: The attributes of the droplets.

(j,k)::Tuple{Int,Int}: The indices of the droplets.

settings::coag_settings{FT}: The coagulation settings.

Droplets.terminal_v_X — Method

terminal_v_X(X::FT) where FT<:AbstractFloat

Compute the terminal velocity of a droplet of volume X